[BUGFIX] Show clipboard if it contains elements
On empty pages no page content is rendered at all. This patch brings the
clipboard back, if it contains elements so you can check those on the
empty pages. Besides the drawFooterHook is shown again on empty pages.
this isn't set in the core itself and extension authors should be able
to choose if the content should be visible on empty pages.

Resolves: #57062
Releases: 6.2
Change-Id: Ie7bd6fda730e801b285d7876a1eff39c3ba1eff0
Reviewed-on: https://review.typo3.org/28603
Reviewed-by: Stefan Froemken
Tested-by: Stefan Froemken
Reviewed-by: Frans Saris
Tested-by: Frans Saris