[BUGFIX] Ensure backwards compatibility in element browser
Hooks that used the element browser used the JavaScript
function jumpToUrl with relative URLs.

After refactoring the element browser, it was necessary
to compute the absolute URL.

To stay backwards compatible we catch this case
in JavaScript and rewrite the URL according
to the new requirements.

By doing so, we can simplify our code again.

The method getThisScript might still be needed in
Hooks for other purposes, so we make it public.

Resolves: #56721
Releases: 6.2
Change-Id: Ieed6d037b428c5c15d18905341ff46dece1d67a7
Reviewed-on: https://review.typo3.org/28618
Reviewed-by: Ernesto Baschny
Reviewed-by: Markus Klein
Tested-by: Markus Klein
Reviewed-by: Stefan Neufeind
Reviewed-by: Wouter Wolters
Reviewed-by: Helmut Hummel
Tested-by: Helmut Hummel