[FEATURE] Improve usability of field selector in record listview
The field selector for the record listview and the element
browser is currently a select field (with multiple options)
and a user must use the control key to select multiple fields.
This is not very user friendly nowadays.

This patch replaces the select field with a list of checkboxes
like used in other places in TYPO3 backend, so a user can
select multiple fields without using the control key. Also the
default column is selected by default and the checkbox is
disabled, because you can't deselect the default column.

Resolves: #62323
Related: #5370
Releases: master
Change-Id: I2a60a0c87766f5b4ea0854d3d8063197c32feea1
Reviewed-on: http://review.typo3.org/33386
Reviewed-by: Jan Helke <typo3@helke.de>
Tested-by: Jan Helke <typo3@helke.de>
Tested-by: Daniel Goerz <ervaude@gmail.com>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>