[BUGFIX] Fix handling of TCA allowed flag in ElementBrowser
If the 'allowed' flag in the element browser contains a '*', which
means any type of file, no file is displayed. This is because a filter
is added that only files are displayed that have a '*' as file ext.

The filter should not be added if the allowed value is '*' or empty.

Resolves: #54504
Documentation: #64609
Releases: master, 6.2
Change-Id: I941566dd51284ecaaa92b42e6e59988c884d91c1
Reviewed-on: http://review.typo3.org/36392
Reviewed-by: Mathias Schreiber <mathias.schreiber@wmdb.de>
Tested-by: Mathias Schreiber <mathias.schreiber@wmdb.de>
Reviewed-by: Philipp Gampe <philipp.gampe@typo3.org>
Tested-by: Philipp Gampe <philipp.gampe@typo3.org>