[BUGFIX] Element browser: correct the display of current link to file
When editing a file link with the element browser wizard,
it is possible to navigate to other tabs. In tabs other than file or
folder, the current link to the file was shown as a simple id rather
than as a path to the file, like in the file or folder tabs.

Releases: master
Resolves: #65012
Change-Id: I79845506b0c5cdb2e956b56b6b505e581df446c7
Reviewed-on: http://review.typo3.org/36842
Reviewed-by: Michael Oehlhof <typo3@oehlhof.de>
Tested-by: Michael Oehlhof <typo3@oehlhof.de>
Reviewed-by: Markus Klein <klein.t3@reelworx.at>
Tested-by: Markus Klein <klein.t3@reelworx.at>