[!!!][FEATURE] ElementBrowser refactoring, LinkBrowser API - Part 1
This change refactors the ElementBrowser and splits it into
multiple dedicated classes for each mode.

The former modes for file, folder and record selection are now
placed in dedicated *Browser classes.
The former mode "wizard", which defines its actual functionality
via the "act" variable, is a separate controller now, which
provides the API to be exensible.

The integration of the various trees and lists is solved by
a dedicated LinkParameterProviderInterface. Any class can
provides link parameters now by implementing this interface.

Other than that:
 * Lots of unused code is removed
 * All JS is moved into requireJS modules
 * Tree code is simplified by properly using constructors

Part 2 will:
 * add Breaking and Feature documentation
 * finalize RTE re-integration
 * finally delete unused classes

Resolves: #66369
Releases: master
Change-Id: I0a28663ce4c91c2405abc9e2a13063699c6bb231
Reviewed-on: http://review.typo3.org/42951
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>