[BUGFIX] Catch some FAL exceptions for missing driver
If a driver of a storage isn't available anymore (e.g. because an
extension was deactivated), an exception is thrown and the backend
stops working on multiple sites.

This patch catches the InvalidArgumentException which is thrown by the
DriverRegistry and ensures a proper working mode. The exception is
caught when:

- editing a record with inline record to a file of that storage
- editing a record with image manipulation of a file of that storage
- opening the element browser where the last folder was within that storage

Resolves: #71296
Releases: master
Change-Id: Ide83d411ab6c0e1850ff9cb9eec894753e73727f
Reviewed-on: https://review.typo3.org/44523
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Alexander Opitz <opitz.alexander@googlemail.com>
Tested-by: Alexander Opitz <opitz.alexander@googlemail.com>