[BUGFIX] JavaScript error on selecting a folder in element browser
If the TCA type group/folder is used, FormEngine renders a field to
select a particular folder. If a different folder is selected in the
tree in the element browser, the initial parameters are not passed
along which leads to the situation, that the target field name is
not know anymore.

Resolves: #72242
Releases: master
Change-Id: I5745f0c022cf63ab9e9de76ee76119dfbd801bea
Reviewed-on: https://review.typo3.org/45382
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Felix Nagel <info@felixnagel.com>
Tested-by: Felix Nagel <info@felixnagel.com>
Reviewed-by: Andreas Fernandez <typo3@scripting-base.de>
Tested-by: Andreas Fernandez <typo3@scripting-base.de>