[TASK] Deprecate BackendUtility::getRecordRaw
Since we follow the principle of "prepared statement", the method
getRecordRaw will break this behaviour. Within the "where" parameter
of the function it is possible to pass a malformed query part.

Thus we should remove it, as the queryBuilder should be used everywhere
in the future to increase visiblity of which query will be processed at
a concrete point of code and to force the concept of prepared statements.

Resolves: #80317
Releases: master
Change-Id: If0028bf897ddee4517228a9e399390fe7266215e
Reviewed-on: https://review.typo3.org/52075
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>