[BUGFIX] Load PageTSconfig in localconf for LinkHandler
The pageTSconfig for LinkHandler is only loaded in ext_tables.php
which means it is only evaluated in the Frontend when a Backend User
is logged in.

However the functionality did work before, if a compiled pageTS was
available in the cache. This could have happened that the pageTS cache
was filled while a BE user was logged in, so it worked (but only in
random cases).

Resolves: #81057
Releases: master, 8.7
Change-Id: I2a2462799a492af2dac90778c51c29cc5c11a782
Reviewed-on: https://review.typo3.org/52646
Reviewed-by: Marc Bastian Heinrichs <typo3@mbh-software.de>
Tested-by: Marc Bastian Heinrichs <typo3@mbh-software.de>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>