[BUGFIX] Do not recalculate PIDs for every query in BE search
The backend search "Show all" link recalculated allowed PIDs for
the current user for all search operations resulting in an
exponentially increasing amount of sql queries depending on the
number of pages in an installation.

As the list of allowed pages for the user does not change during
the request, it is enough to calculate it once.

Change-Id: I9262a5d8425d04cb4ff1e35b2c3cc0ec96b8fa53
Resolves: #82083
Releases: master, 8.7
Reviewed-on: https://review.typo3.org/53930
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Joerg Boesche <typo3@joergboesche.de>
Tested-by: Joerg Boesche <typo3@joergboesche.de>
Reviewed-by: Helmut Hummel <typo3@helhum.io>
Reviewed-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>
Tested-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>