[BUGFIX] Remove legacy code for version extension
The "Display version" module of ext:version has been
moved to ext:compatibility7 with v8. The list module links
to this view but the link is dead if compat7 is not loaded.
The v8 version of the patch now checks for "loaded version,
loaded compat7, but not loaded workspaces" to render the link,
in master, this button is removed entirely

Resolves: #82394
Related: #79942
Releases: master, 8.7
Change-Id: I4e32a63b144b658f84362e25bda1cc312247367e
Reviewed-on: https://review.typo3.org/54044
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>