[TASK] Protect RecordListController
Protect properties and methods of RecordListController. The two
hooks 'drawHeaderHook' and 'drawFooterHook' now receive an instance
of the PSR-7 request, handing over the parent object is deprecated
and will be removed in v10.
The sys_notes hooks are adapted accordingly and used as a migration
example in the .rst file.

Resolves: #86198
Releases: master
Change-Id: Idee82d2e675eefaee1ead7435bc17cfc3a48356b
Reviewed-on: https://review.typo3.org/58231
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>