[FEATURE] API to search for files including meta data based on folder
Instead of globally selecting all files with matching
meta data, we now base each search on the selected folder.

To do so, search is now completely based on the persisted
files in sys_file (and their corresponding meta data
in sys_file_metadata).

Additionally we properly evaluate search fields from TCA
so that we now search in all fields defined for sys_file
and sys_file_metadata table.

To achieve that, a new capability "CAPABILITY_HIERARCHICAL_IDENTIFIERS"
is introduced, which drivers can set, that build identifiers
that reflect the directory structure.
For such drivers, the search can be optimized by using
like queries on identifiers, instead of recursively traversing folders,
which can be an expensive operation especially for drivers
handling a remote file system.

Resolves: #87610
Releases: master, 9.5
Change-Id: Ia132465437827b2fdb56004eb73348ce4a05b336
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/60685
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>