[BUGFIX] Allow adding content without page edit permissions in record list
If a backend usergroup has no permission to edit page properties, the
wizard for content element creation is not rendered properly. This is due
to a missing import of the `PageActions` module.

This patch now always loads the module and makes its callback depending
on proper permissions.

Resolves: #88532
Related: #87687
Releases: master, 9.5
Change-Id: Ie3acaf8fbf9bdeee425eae4b4d35cb2eb80b4d1d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/60907
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Markus Klein <markus.klein@typo3.org>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>