[BUGFIX] Keep priority of SQL operators in DatabaseRecordList
DatabaseRecordList adds an additionalWhereClause with an `OR` which
may be extended by a hook. Usually, the hook is used to add additional
clauses. As the `OR` part is not wrapped in braces, this may result in
wrong priorities in the resulting query.

As there might be pre-existing work-arounds handling the situation,
the fix is done only on master.

Resolves: #82283
Releases: master
Change-Id: Id60d02cf671db2fffc2635764ebda809ae7cb1af
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/63620
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>