[BUGFIX] Keep order of selection on importing assets
If assets get imported via mass selection in the Element Browser, the
order of the items is now kept. Since this requires interframe
communication, the Inline Control Container now sends a postMessage to
the Element Browser.

When the import process starts, the first asset of the selection gets
imported. After the import, the postMessage is intercepted and the next
asset get imported, until the queue is empty. As this could be a rather
lenghty process, a progress bar is rendered at the top of the file
table.

To get a more fitting naming, all previously used action names used in
postMessages have been renamed from `typo3:foreignRelation:inserted` to
`typo3:foreignRelation:insert`.

As a drive-by fix, the server is not spammed with n request anymore due
to queueing.

Resolves: #86362
Resolves: #59341
Releases: master
Change-Id: Ib9e3ee8d943582874a0a73632968f8660b169b42
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/63838
Tested-by: Susanne Moog <look@susi.dev>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>
Reviewed-by: Susanne Moog <look@susi.dev>
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>