[TASK] DatabaseRecordList: Streamline limit handling
The actual limit, used in the getTable(), depends on various
configurations, current view (default view, single table view)
and current action (view, search or csv export).

To get a better overview, the patch moves limit handling to the
getTable() method, so limit handling in start() and generateList()
can be removed. The properties itemsLimitPerTable
and itemsLimitSingleTable are removed as well.

The configured limit (excluding search and CSV export) is now
forced to be in a range from 5 to 10000.

For translated pages, the pages configuration is used to set the limit.

The limit is set by the following chain:
[default view with multiple tables; single table view]

- default value [20; 100]
- if page ts config is set [itemsLimitPerTable; itemsLimitSingleTable]
- if tca configuration is set [maxDBListItems; maxSingleDBListItems]
- a forced range from 5 to 10000
- if limit is set in the search
- no limit for a csv export

Releases: master
Resolves: #92103
Change-Id: I823c039aa063df8f99e08e190990c5e34466c253
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65461
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>