[BUGFIX] Workspace delete placeholder handling in list module
When a record that exists in live is deleted in a workspace, a
new 'delete placeholder' record is created in the workspace. This
record is used to mark the live workspace record as 'to be deleted'
when the workspace is published.

Those delete placeholder are not shown in the page tree and not
in the page module. The page / element is simply gone. Only the
list module renders those records.

This is confusing usability wise: A user clicks on 'delete' in
list module, reloads the module, and the record is back again.

Additionally, there is an even more confusing mechanism: The 'waste'
icon is substituted by a different icon called 'restore'. 'restore'
however does not actually restore changes that may have been done
to the record in the workspace before, instead it simply sets the
delete placeholder record to deleted, which makes the underlying
live record 'shine through' in the workspace again.

Interacting with delete placeholder records is also useless,
for instance edited content of delete placeholder records is
always gone in the end. So most action icons don't make sense.

Looking at the lifecycle of workspace records, there are only
two final results: A change is either discarded, or is published
to live. This is managed exclusively in the workspace module.
Having the third option 'restore' in the list module - that is
basically the same as discarding a delete placeholder in workspace
module - is very confusing.

The best option would be to not render delete placeholder rows
in the list module at all. Unfortunately this can't be done due
to the workspace overlay mechanics: Especially the count and
limit / pagination handling in list module is done on db level before
the module knows that certain rows shouldn't be rendered. We need to
have some serious refactoring before this can be solved.

For now, the solution is to still render delete placeholder records
in list module, to 'strike through' their title (similar to workspace
module) and to disable all control icons and the click menu on them.
This way the final fate of a delete placeholder can be decided only
in the workspace module - just like all other workspace records.

Change-Id: Id8a95c41f2bbe90ea68f40e376a4d36d3a2019b8
Resolves: #90679
Resolves: #52700
Releases: master, 10.4
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65495
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Daniel Windloff
Tested-by: Riccardo De Contardi <erredeco@gmail.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Daniel Windloff
Reviewed-by: Björn Jacob <bjoern.jacob@tritum.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>