[BUGFIX] DatabaseRecordList: Field selection in single table view
The following items have been removed as they haven't any impact:

* `_LOCALIZATION_`
* `_CLIPBOARD_`
* `_CONTROL_`

`_LOCALIZATION_` and `_CLIPBOARD_` are always filtered out.
`_CONTROL_` is always added, if not selected in the field selection.

Furthermore, the `ctrl` values `tstamp` and `crdate` could be added
by any user regardless of the column name
(respecting the configuration in TCA).

Finally, the column showing the references is added, if selected.

Releases: master, 10.4
Resolves: #80396
Change-Id: Ie809a0e6f76e62166dc2d6f9cecfb3c68750e70e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65501
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>