[TASK] DatabaseRecordList: Streamline `fieldArray`
The `fieldArray` related configuration has been moved to the `getTable`
method.
The configuration handling for the `fieldArray` has been changed, in
order to avoid later unset on the `fieldArray`.

The `fieldArray` is only in the `getTable` and subordinated render
methods.
Therefore, the implicit configuration transfer like `fieldArray[0]`
for the label column has been replaced.

Releases: master
Resolves: #92154
Change-Id: I04287e90e8cffc2ab768bd5399a8155f823fdfe1
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65534
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>