[FEATURE] Introduce Bootstrap v5 for TYPO3 Backend
This changes removes the frontend framework
Bootstrap 3, and adds Bootstrap 5 beta 1 (we
expect Bootstrap 5 final by the time we release TYPO3 v11 LTS).

Bootstrap v3 is not supported by the Bootstrap
team any longer, so an update is critical for TYPO3 Core.

Bootstrap v5 adds a few accessibility improvements
as well as flexbox for rendering
containers and grids throughout TYPO3 Backend.

All JS components are not bound to jQuery anymore,
and have been reworked.

A lot of HTML/CSS changes happened, which we
slowly migrate (and not in a huge change)
to TYPO3's templates, in order to keep
this change managable.

A legacy CSS/SCSS file is added to
keep some backwards-compatibility classes
to ease the migration for extension developers
who have built their own backend modules.

Key features of Bootstrap 5:
* "rem" instead of "px" is used by default
* CSS variables are introduced
* Improved bootstrap focus outline styling (buttons / inputs / links)
* Simplified grid functionality
* use new button color mixin to increase contrast:
  Primary, Success and Warning Button color is now dark instead of white

EXT:styleguide was used as a basis for
upgrading to keep compatibility as much
as possible, but more changes will be coming
in the next few minor releases.

Resolves: #92616
Releases: master
Change-Id: Iec989f39649b5460b055ec879199faf38e424f2b
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/66247
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benjamin Franzke <bfr@qbus.de>
Tested-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>