[TASK] Extract CSV output into separate functionality
The CSV output of the database record list (= list module)
is one of the only parts that cannot be modified via middlewares
as its rendering directly exits PHP.

This makes it impossible to test this functionality.

This change moves the CsvExport into its own class CsvExportRecordList
which accumulates the records from the database records list.

The existing hooks are migrated and kept as is (due to the
Composition pattern).

The class is a composition for RecordList to extract
the querying logic into a separate code.

The Controller class is responsible of handling the CSV
response via the PSR-17 Response Factory.

Resolves: #88713
Releases: master
Change-Id: I6880998cd044a1b265bbd0347cfe084911998b34
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/68664
Tested-by: core-ci <typo3@b13.com>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>