[TASK] Move doc header generation out of DatabaseRecordList
This change primarily moves the ModuleTemplate / DocHeader
generation out of the DatabaseRecordList class into the actual
controller where ModuleTemplate is used.

The main controller method is also split up in smaller chunks
(such as Clipboard initialization).

In addition, several methods and properties are marked
as protected if they are only used internally within DatabaseRecordList.

Further steps in this area are to separate code from DatabaseRecordList
which is not bound to the record listing.

Resolves: #93863
Releases: master
Change-Id: I901b0acebaec8f53b04426cad93002a3ec74195e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/68712
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>