[TASK] Remove legacy tree implementations
Various PHP classes, which were previously a specific implementation
for various Link Picker / Element Browser renderings for trees
are removed.

In addition, the so-called "LegacyTree" is removed along with
all entry points, also for the "navFrameScript" of the iframe inside
the File main module.

Modules who previously registered to the iFrame navigation
component are now moved to the new file storage component.

Since the related code was marked as internal, the files are removed
without substitution.

Resolves: #93944
Releases: master
Change-Id: Ib42c5b32d44c8a697e250f178280f743b0d7e1da
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/68799
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>