[TASK] Set proper title of window in backend
With the introduction of the new backend module web component router,
the title of the backend windows will be set to the title of the
main iframe. Most of the modules didn't provide a proper name though.

For most modules we have a proper name now which will show up in the
title of the backend window, if no title is propagated by the module,
the backend router will fallback to the default backend title.

As the format of the title is quite "personal". If you are used to have
opened more TYPO3 backend windows, you would like to see which
installation you have open. If you only work in one backend, you
might want to see on which module you are currently working. It is
possible to set the order of the title of the backend within your
user settings now. By default it will be title - siteName [version]

Resolves: #94182
Releases: master
Change-Id: I02602650370140217aa252bbd8e29ea4e05d994a
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/69172
Tested-by: core-ci <typo3@b13.com>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Benjamin Franzke <bfr@qbus.de>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>