[BUGFIX] Only show contextual language information in List module
The List module shows the site information next to a language,
which is not needed in case when listing languages of a single
page (where the PID is given anyways).

In addition, for pages the context is the current uid
and not the "pid" itself.

The change adapts the list, making the overview much
more practical when dealing with multiple sites.

Resolves: #94171
Releases: master, 10.4
Change-Id: I17b3b83ceca98f88556dc5caad6d84c1b96bf9c9
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/69272
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>