[FEATURE] Customize selectable columns in List module per table
In prior TYPO3 versions, it was possible to customize the list of fields
to be shown in the list view. This was only possible if the clipboard
was active OR the user was in "single table view", and displayed
in a box at the bottom of the module without any context.

The new field selector is now right at the top of the database table -
and per table - thus, always active and in context with the table.

Via UserTSconfig it is possible to explicitly disable the functionality
for users/groups for all tables or just a specific table.

Example: Disable it for all tables except for sys_category:

   mod.web_list.displayColumnSelector = 0
   mod.web_list.table.sys_category.displayColumnSelector = 1

Resolves: #94218
Releases: master
Change-Id: Ia6db3a18fea5d0cf96fa32611535048f6e5c6199
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/69262
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>