[BUGFIX] Do not render clipboard actions for page translations
Recordlist features the clipboard functionality. When activated,
multiple clipboard actions, such as "copy" or "edit marked", are
displayed in each tables' header.

Since translated records can never be selected, this led to
exceptions and JavaScript errors when using such action
on the "Page Translations" table.

This is now fixed by not rendering those actions for this
special table.

Resolves: #94296
Releases: master, 10.4
Change-Id: I47459cf1d82d90f634a4f8902ea04c958cbd1bdc
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/69427
Tested-by: core-ci <typo3@b13.com>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Oliver Bartsch <bo@cedev.de>