[BUGFIX] Fix "hideTable" check in RecordListController
Since the method has to return true in case the page
translations should be displayed, the TCA "hideTable"
check must be negated. This would otherwise lead to
page translations not being displayed if the "hideTable"
key is not defined in pages TCA. Furthermore, the TSconfig
option "hideTable" also contains the `table.` key.

Note: Only master is affected as backport was not yet
merged and will therefore contain the correct check.

Resolves: #94398
Related: #94386
Releases: master
Change-Id: I96b2676b3acde6bfc3436fd16fc31ca76ba3d185
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/69539
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>