[FEATURE] Improve show columns selection functionality
In #94218, the show columns selection was already improved
by relocating it to each tables' header, making it always
available (not only in single-table view).

Since the used dropdown leads to confusion and visibility
drawbacks for records, having long labels, the selector
now lives in a modal. Besides the columns to select,
the new modal contains some selector options, such as
"select all" or "toggle selection". Those options are
fixed at the top, making them always visible, even for
records with a lot of columns.

Furthermore the checkboxes now use a more suitable
"check" icon, the columns are sorted lexically and
also management fields (e.g. `sorting`) are now
displayed with a human-readable label. This label
is now also used in the table header.

Besides the UX, this patch also brings some technical
improvements. The whole column selection logic is moved
into a dedicated controller, which may allows reuse in
the future (e.g. for the filelist module). Also the
JavaScript part is extracted from the Recordlist JS-module
into a dedicated web component.

Finally some cleanup of the existing code is done, making
it more readable and more efficient.

Resolves: #94474
Releases: master
Change-Id: I0f2f8711ee4b40c9e29e633b9fe790dcacae8bb5
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/69691
Tested-by: core-ci <typo3@b13.com>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>