[TASK] Streamline File List and Clipboard Selection Options
This change moves the "Display Thumbnails" selection above the
File List table, which then unifies this selection with the File Selector
(IRRE File Selection).

In addition, the only leftover option (Show clipboard) is now also
a switch checkbox, and also applied in the view for the RecordList.

All docheader buttons in the File Listing now show the appropriate
text to make it easier to understand what the user clicks.

Resolves: #94517
Releases: master
Change-Id: I4ef02ba4aba008ae86b7452695a5590f2846a296
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/69770
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>