[TASK] Always show "normal" clipboard actions in secondary menu
This patch moves the "normal" clipboard actions, such
as copy and cut, into the secondary dropdown in
recordlist and filelist. This means, when in "normal"
clipboard mode, the actions are accessible no matter
if the clipboard itself is shown or not.

The clipboard related actions are divided by a horizontal
line in the dropdown.

Moving them into the dropdown also means, no superfluous
"empty" icons are necessary when clipboard state changes,
e.g. from "normal" to "multiple".

Resolves: #94824
Releases: master
Change-Id: I57666fe880cda9283d649a77f3a3b1bd44206e8d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/70455
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>